# 8268427: Improve AlgorithmConstraints:checkAlgorithm performance