@seanjmullan 2021-06-10
I think it would be worthwhile to see if we can take this a step further, and leverage the Constraint objects that are already created and stored in the `Constraints` object in a `HashMap<String, List<Constraint>>`. The key is the algorithm `String`, and the value is a `List` of `Constraint` objects that apply to it. If the algorithm is completely disabled, the `List` would contain one entry `Constrant.DisabledConstraint` (which has a `permits` method that always returns `false`).

This way we could potentially eliminate the List/Set cache entirely and just use the HashMap to check if the algorithm is disabled.