@dongbohe 2021-06-17
The following is the Benchmark results for the new commit:

JMH:
```java
Benchmark                            (algorithm)  Mode  Cnt    Score    Error  Units
AlgorithmConstraintsPermits.permits        SSLv3  avgt    5   47.353 ?  3.193  ns/op
AlgorithmConstraintsPermits.permits          DES  avgt    5   60.307 ?  1.351  ns/op
AlgorithmConstraintsPermits.permits         NULL  avgt    5   59.065 ?  0.728  ns/op
AlgorithmConstraintsPermits.permits       TLS1.3  avgt    5  428.311 ? 16.542  ns/op
```
Tomcat+Jmeter:
```java
summary +  60455 in 00:00:30 = 2016.4/s Avg:   198 Min:   164 Max:   256 Err:     0 (0.00%) Active: 400 Started: 400 Finished: 0
summary = 106982 in 00:00:55 = 1927.6/s Avg:   188 Min:    25 Max:  1194 Err:     0 (0.00%)
summary +  60430 in 00:00:30 = 2014.6/s Avg:   198 Min:   160 Max:   252 Err:     0 (0.00%) Active: 400 Started: 400 Finished: 0
summary = 167412 in 00:01:25 = 1958.1/s Avg:   191 Min:    25 Max:  1194 Err:     0 (0.00%)
summary +  60480 in 00:00:30 = 2014.6/s Avg:   198 Min:   161 Max:   245 Err:     0 (0.00%) Active: 400 Started: 400 Finished: 0
summary = 227892 in 00:01:56 = 1972.8/s Avg:   193 Min:    25 Max:  1194 Err:     0 (0.00%)
summary +  60506 in 00:00:30 = 2016.6/s Avg:   198 Min:   161 Max:   255 Err:     0 (0.00%) Active: 400 Started: 400 Finished: 0
```