@ascarpino 2021-06-18
I will let Xuelei finish his review.  From what I see in the code, the Comparator is used when the algorithm is fully disabled.  The use of keywords later on in the Constraints constructor uses the ConstraintsEntry which I believe is stored with the upper and lower case.  I think the code should be ok.